Make results in tests a list
Previously we could mock one result, and the outcome of an operation
would always be the same. Now we can pass in a list of results, and on
every call, we take one result from the list and use that. To get the
old behavior, wrap the result in "repeat". The list has two advantages:

 * We can supply a finite amount of results, and catch unexpected calls
   this way.

 * We can make different calls return different results.