Use `--force-with-lease` instead of `--force`
Using `--force` in `git push` can be destructive as it does not take
into account the state of the remote. With `--force-with-lease` the push
will fail if the remote is in a unexpected state. It is better to exit
early during the push that to carry on and cause trouble on the
remote.