test setup: specify init branch
in the test setup for the end-to-end test suite, we rely on the
assumption that the local git is configured to use `master` for
the `init.defaultBranch`, causing tests to fail on systems where the
assumption does not hold true.

specifying it explicitly using the `--initial-branch` option prevents
tests from failing due to the runner's git config.